refactor: Move Connection/Profile code in to own module
Moves the connectio/profile code from src/auth/auth.ts
in to src/auth/connection.ts.

This was done to separate the connection code from the
auth client in auth.ts since external code uses connections
but auth may use that external code, causing a circular
dependency
