+++ src/test/codewhisperer/util/authUtil.test.ts
@@ -4,12 +4,13 @@
  */
 
 import * as assert from 'assert'
-import { codewhispererScopes, isSsoConnection } from '../../../auth/auth'
+import { isSsoConnection } from '../../../auth/auth'
 import { AuthUtil } from '../../../codewhisperer/util/authUtil'
 import { getTestWindow } from '../../shared/vscode/window'
 import { SeverityLevel } from '../../shared/vscode/message'
 import { createBuilderIdProfile, createSsoProfile, createTestAuth } from '../../credentials/testUtil'
 import { captureEventOnce } from '../../testUtil'
+import { codewhispererScopes } from '../../../auth/connection'
 
 const enterpriseSsoStartUrl = 'https://enterprise.awsapps.com/start'
