+++ src/test/credentials/auth.test.ts
@@ -5,14 +5,7 @@
 
 import * as assert from 'assert'
 import * as sinon from 'sinon'
-import {
-    AuthNode,
-    Connection,
-    isIamConnection,
-    isSsoConnection,
-    promptForConnection,
-    ssoAccountAccessScopes,
-} from '../../auth/auth'
+import { AuthNode, isIamConnection, isSsoConnection, promptForConnection } from '../../auth/auth'
 import { ToolkitError } from '../../shared/errors'
 import { assertTreeItem } from '../shared/treeview/testUtil'
 import { getTestWindow } from '../shared/vscode/window'
@@ -25,6 +18,7 @@ import { makeTemporaryToolkitFolder } from '../../shared/filesystemUtilities'
 import { SharedCredentialsProviderFactory } from '../../auth/providers/sharedCredentialsProviderFactory'
 import { UserCredentialsUtils } from '../../shared/credentials/userCredentialsUtils'
 import { getCredentialsFilename } from '../../auth/credentials/sharedCredentialsFile'
+import { Connection, ssoAccountAccessScopes } from '../../auth/connection'
 
 const ssoProfile = createSsoProfile()
 const scopedSsoProfile = createSsoProfile({ scopes: ['foo'] })
