+++ src/test/credentials/testUtil.ts
@@ -5,7 +5,7 @@
 
 import * as assert from 'assert'
 import * as vscode from 'vscode'
-import { Auth, Connection, ProfileStore, SsoConnection, SsoProfile } from '../../auth/auth'
+import { Auth } from '../../auth/auth'
 import { CredentialsProviderManager } from '../../auth/providers/credentialsProviderManager'
 import { SsoClient } from '../../auth/sso/clients'
 import { builderIdStartUrl, SsoToken } from '../../auth/sso/model'
@@ -17,6 +17,7 @@ import globals from '../../shared/extensionGlobals'
 import { fromString } from '../../auth/providers/credentials'
 import { mergeAndValidateSections, parseIni } from '../../auth/credentials/sharedCredentials'
 import { SharedCredentialsProvider } from '../../auth/providers/sharedCredentialsProvider'
+import { Connection, ProfileStore, SsoConnection, SsoProfile } from '../../auth/connection'
 
 export function createSsoProfile(props?: Partial<Omit<SsoProfile, 'type'>>): SsoProfile {
     return {
